+++ package.json
@@ -38,6 +38,7 @@
     "react-chartjs-2": "^5.3.0",
     "react-device-detect": "^2.2.3",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.55.0",
     "react-lottie": "^1.2.10",
     "react-redux": "^9.2.0",
     "react-router-dom": "^7.1.3",
@@ -46,7 +47,8 @@
     "redux-persist": "^6.0.0",
     "redux-thunk": "^3.1.0",
     "swr": "^2.3.3",
-    "uuid": "^11.0.5"
+    "uuid": "^11.0.5",
+    "zod": "^3.24.2"
   },
   "devDependencies": {
     "@chromatic-com/storybook": "^3",
