+++ src/app/Debug/SignIn.jsx
@@ -1,216 +1,70 @@
 import { useState, useEffect } from "react";
 
 import { Container, Grid2, Typography, Button, Paper, IconButton, Box, Tooltip } from "@mui/material";
-import { Snackbar, Alert } from "@mui/material";
-
 import { GlobalStyles } from "@config";
-
-import { BpForm, BpLoading } from "@components";
-import { useForm, useToggle } from "@hooks";
 
 import { Amplify } from "@libs/auth";
 
+import { fetchAuthSession, getCurrentUser, fetchUserAttributes } from "@aws-amplify/auth";
 
-import { fetchAuthSession, getCurrentUser } from "@aws-amplify/auth";
-const { handleSignIn, handleSignUp, handleConfirmSignUp, handleSignOut, handleResetPassword, handleConfirmResetPassword, isAuthenticated, handleResendSignUpCode } = Amplify;
+import { Authenticator } from '@aws-amplify/ui-react';
+import '@aws-amplify/ui-react/styles.css';
 
-const template = {
-    login: {
-        key: "login",
-        field: [
-            {
-                "name": "username",
-                "type": "email"
-            },
-            {
-                "name": "password",
-                "type": "password"
-            }
-        ]
+import { styled } from "@mui/material";
+
+const BpContainer = styled(Container)(({ theme }) => ({
+    height: 'calc((1 - var(--template-frame-height, 0)) * 100dvh)',
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+    minHeight: '100%',
+    '&::before': {
+        content: '""',
+        display: 'block',
+        position: 'absolute',
+        zIndex: -1,
+        inset: 0,
+        backgroundImage: 'linear-gradient(180deg, rgba(80,150,209, 0.2) 0%, rgba(255,255,255,1) 75%)',
+        backgroundRepeat: 'no-repeat',
+        ...theme.applyStyles('dark', {
+            backgroundImage: 'radial-gradient(at 50% 50%, hsla(210, 100%, 16%, 0.5), hsl(220, 30%, 5%))',
+        }),
     },
-    signUp: {
-        key: "sign_up",
-        field: [
-            {
-                "name": "username",
-                "type": "email"
-            },
-            {
-                "name": "password",
-                "type": "password"
-            },
-            {
-                "name": "confirm_password",
-                "type": "password"
-            }
-        ]
-    },
-    otpCode: {
-        key: "otp_code",
-        field: [
-            {
-                "name": "username",
-                "type": "email"
-            },
-            {
-                "name": "code",
-                "type": "text"
-            }
-        ]
-    }
-}
+}));
 
-function TxtButton(props) {
-
-    const { onClick = () => { }, children = (<></>) } = props;
-
+function RedirectToFirstTimeLogin(props) {
     return (
-        <Typography
-            sx={{
-                background: "none",
-                border: "none",
-                cursor: "pointer",
-                "&:hover": { textDecoration: "underline" },
-            }}
-            onClick={onClick}
-        >
-            {children}
-        </Typography>
-    );
+        <Grid2>
+            <Typography>Redirecting to First Time Login...</Typography>
+        </Grid2>
+    )
 }
+
+import { ColorModeIconDropdown } from '@components';
 
 function Index(props) {
 
-    const { key: lKey, data: lData, field: lField, updateDataHtml: updateLData, resetData: resetLData } = useForm(template.login);
-    const { key: sKey, data: sData, field: sField, updateDataHtml: updateSData, resetData: resetSData } = useForm(template.signUp);
-    const { key: oKey, data: oData, field: oField, updateDataHtml: updateOData, resetData: resetOData } = useForm(template.otpCode);
-
-    const { flag: resendFlag, toggle: toggleResend } = useToggle(false);
-    const { flag: userFlag, open: setUserTrue, close: setUserFalse, toggle: toggleUser } = useToggle(false);
-
-    const [session, setSession] = useState({});
-    const [signUpSession, setSignUpSession] = useState({});
-
-    const [loading, setLoading] = useState(false);
-
-    useEffect(() => {
-        console.log(session);
-    }, [session])
-
-    const onLogin = () => {
-        handleSignIn(lData)
+    const onDebug = () => {
+        fetchUserAttributes()
             .then((res) => {
-                alert("User has signed in!");
-                setUserTrue();
-
-                // Store this Data inside Redux, with Last Session Date
-                setSession(_ => res);
+                console.log("session", res);
             })
-            .catch(err => {
-                alert(err);
-            })
-    }
-    const onLogOut = () => {
-        handleSignOut()
-            .then(() => {
-                alert("User has signed out!");
-                setUserFalse();
-            })
-    }
-
-    const onSignUp = () => {
-        handleSignUp(sData)
-            .then((res) => {
-                alert("User has signed up! Please Check Email for Activation Code!");
-
-                setSignUpSession(_ => res);
-            })
-            .catch(err => {
-                alert(err);
-            })
-    }
-
-    const onConfirmSignUp = () => {
-        handleConfirmSignUp(oData)
-            .then(() => {
-                alert("User has confirmed sign up!");
-            })
-            .catch(err => {
-                alert(err);
-            })
-    }
-
-    const onResendSignUpCode = () => {
-        toggleResend();
+            .catch((e) => {
+                console.log("error", e);
+            });
     }
 
     return (
-        <>
-            <BpLoading loading={loading} />
-
-            <Container maxWidth={"xl"} sx={{ padding: 2 }}>
-                <Box>
-                    <Typography variant={"h2"}>Amplify Cognito</Typography>
-                </Box>
-                <Box sx={{ mt: 2, ...GlobalStyles.bordered }}>
-                    <BpForm
-                        key={lKey} idx={lKey}
-                        data={lData} field={lField}
-                        onUpdate={updateLData}
-                        hasLabel={true} size={{ xs: 1, sm: 1 }} />
-                    <Grid2 container spacing={1} sx={{ mt: 2 }}>
-                        <Button variant={"contained"} onClick={onLogin}>Login</Button>
-                        <Button variant={"contained"} onClick={onLogOut}>Logout</Button>
-                    </Grid2>
-                </Box>
-                <Box sx={{ mt: 2, ...GlobalStyles.bordered }}>
-                    <BpForm
-                        key={sKey} idx={sKey}
-                        data={sData} field={sField}
-                        onUpdate={updateSData}
-                        hasLabel={true} size={{ xs: 1, sm: 1 }} />
-                    <Box sx={{ mt: 2 }}>
-                        <Button variant={"contained"} onClick={onSignUp}>Sign Up</Button>
-                    </Box>
-                </Box>
-                <Box sx={{ mt: 2, ...GlobalStyles.bordered }}>
-                    <BpForm
-                        key={oKey} idx={oKey}
-                        data={oData} field={oField}
-                        onUpdate={updateOData}
-                        hasLabel={true} size={{ xs: 1, sm: 1 }} />
-                    <Box sx={{ mt: 1 }}>
-                        <TxtButton onClick={onResendSignUpCode}>Resend Code</TxtButton>
-                    </Box>
-                    <Snackbar open={resendFlag} autoHideDuration={3000}
-                        onClose={toggleResend}
-                        anchorOrigin={{
-                            vertical: "bottom",
-                            horizontal: "center",
-                        }}>
-                        <Alert onClose={toggleResend} severity="success" variant="filled">
-                            Successfully Resend OTP Code!
-                        </Alert>
-                    </Snackbar>
-
-                    <Box sx={{ mt: 2 }}>
-                        <Button variant={"contained"} onClick={onConfirmSignUp}>Confirm OTP</Button>
-                    </Box>
-                </Box>
-                <Box sx={{ mt: 1 }}>
-                    {
-                        (userFlag) ? (
-                            <Typography variant={"h4"}>Welcome, User</Typography>
-                        ) : (
-                            <Typography variant={"h4"}>Not Signed In</Typography>
-                        )
-                    }
-
-                    <Typography variant={"h4"}>{JSON.stringify(session)}</Typography>
-                    <Typography variant={"h4"}>{JSON.stringify(signUpSession)}</Typography>
-                </Box>
-            </Container>
-        </>
+        <BpContainer maxWidth={"xl"}>
+            <ColorModeIconDropdown sx={{ position: 'fixed', top: '1rem', right: '1rem' }} />
+            <Authenticator
+                socialProviders={['google']}
+                signUpAttributes={['email']}>
+                {({ signOut, user }) => (
+                    <RedirectToFirstTimeLogin />
+                )}
+            </Authenticator>
+        </BpContainer>
     )
 }
