--- src/components/ui/Form/Item/FormColor.jsx
@@ -1,14 +1,0 @@
-
-import { TextField, Typography, Grid2, FormControlLabel, Switch } from "@mui/material";
-
-import { MuiColorInput } from 'mui-color-input';
-
-const Index = (props = {}) => {
-    const { value = "", onChange = () => { }, sx = {} } = props;
-
-    return (
-        <MuiColorInput format="hex" value={value} onChange={onChange} style={sx} />
-    )
-}
-
-export default Index;
