+++ src/components/ui/Form/Item/FormItem.jsx
@@ -7,15 +7,7 @@ import BasicFormItem from "./BasicFormItem";
 
 const FormItem = (props = {}) => {
 
-    const { hasLabel = false } = props;
-
-    const { required = false, name = "", size = {} } = props;
-
-    const style = {
-        txtLabel: {
-            width: { xs: "120px", sm: "135px" },
-        }
-    }
+    const { size = {} } = props;
 
     const _size = {
         xs: 12, 
@@ -23,13 +15,9 @@ const FormItem = (props = {}) => {
         ...size,
     }
 
-    // Parameters
-    const placeholder = clsUtility.capitalize(name);
-    const lbl = hasLabel ? "" : placeholder;
     return (
         <Grid2 item size={_size} sx={{ display: "flex", alignItems: "center"}}>
-            {hasLabel ? (<Typography sx={style.txtLabel}>{placeholder}{required ? "*" : ""}</Typography>) : (<></>)}
-            <BasicFormItem label={lbl} placeholder={placeholder} {...props} />
+            <BasicFormItem {...props} />
         </Grid2>
     );
 }
