+++ src/config/models/Sample/index.js
@@ -1,0 +1,68 @@
+
+import { SampleData } from "@config";
+
+const obj = {
+    key: "Incident",
+    field: [
+        {
+            "name": "title",
+            "type": "text"
+        },
+        {
+            "name": "scammer_type",
+            "type": "dropdown",
+            "selection": [
+                {
+                    "name": "Seller",
+                    "value": "Seller"
+                },
+                {
+                    "name": "Buyer",
+                    "value": "Buyer"
+                }
+            ]
+        },
+        {
+            "name": "subtitle",
+            "type": "textarea"
+        },
+        {
+            "name": "description",
+            "type": "textarea"
+        },
+        {
+            "name": "background",
+            "type": "color"
+        },
+        {
+            "name": "post_date",
+            "type": "date"
+        },
+        {
+            "name": "email",
+            "type": "email"
+        },
+        {
+            "name": "password",
+            "type": "password"
+        },
+        {
+            "name": "quantity",
+            "type": "int"
+        },
+        {
+            "name": "total_amount",
+            "type": "decimal"
+        },
+        {
+            "name": "profile",
+            "type": "image"
+        },
+        {
+            "name": "file",
+            "type": "file"
+        },
+    ]
+}
+
+export default obj;
