+++ src/config/models/index.js
@@ -7,12 +7,15 @@ import Incident from "./Incident";
 import IpSeries from "./IpSeries";
 import Account from "./Account";
 
+import Sample from "./Sample";
+
 export {
     Product,
     Scammer,
     ScammerAttr,
     Incident,
     IpSeries,
-    Account
+    Account,
+    Sample
 }
